// packages/discord/src/operations/listMyPrivateArchivedThreads.ts
export const ListMyPrivateArchivedThreadsInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    channel_id: Schema.String.pipe(T.PathParam()),
    before: Schema.optional(Schema.String),
    limit: Schema.optional(Schema.Number),
  }).pipe(
    T.Http({
      method: "GET",
      path: "/channels/{channel_id}/users/@me/threads/archived/private",
    }),
  );
export type ListMyPrivateArchivedThreadsInput =
  typeof ListMyPrivateArchivedThreadsInput.Type;
// ...
export type ListMyPrivateArchivedThreadsOutput =
  typeof ListMyPrivateArchivedThreadsOutput.Type;

// The operation
export const listMyPrivateArchivedThreads =
  /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
    inputSchema: ListMyPrivateArchivedThreadsInput,
    outputSchema: ListMyPrivateArchivedThreadsOutput,
    errors: [BadRequest, Forbidden, NotFound] as const,
  }));